+++ src/components/Checkout/CheckoutOrder.vue
@@ -240,7 +240,7 @@ export default {
 
 
         //  går till en sida som visar att order är skickad ?
-                   this.$router.push("/shop");
+                   this.$router.push("/checkout/checkoutOrderOk");
         // eller visa en modal ?
 
       } else {
