+++ src/components/Checkout/CheckoutOrderOk.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="container pt-5">
+    <div class="row">
+      <div class="col">
+        <h4>Your order has been submitted</h4>
+      </div>
+    </div>
+  </div>
+</template>
+
+<script>
+export default {
+   name: "CheckoutOrderOk",
+};
+</script>
+
+<style>
+</style>
